@@ -1,18 +1,19 @@
 name: network_manager
 description: Dart package which manages data connection.
-version: 1.0.0
+version: 2.0.0
 # homepage: https://www.example.com
-
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  data_connection_checker: ^0.3.4
+  data_connection_checker:
+    git:
+      url: git://github.com/uSlashVlad/data_connection_checker.git
   equatable: ^2.0.0
-  meta: ^1.1.8
-  path: ^1.7.0
+  meta: ^1.3.0
+  path: ^1.8.0
 
 dev_dependencies:
-  mockito: ^5.0.3
-  test: ^1.14.4
+  mocktail: ^0.1.1
+  test: ^1.16.8
   very_good_analysis: ^2.0.0
