@@ -1,4 +1,4 @@
-import 'package:mockito/mockito.dart';
+import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
 
 import 'package:network_manager/network_manager.dart';
@@ -6,8 +6,8 @@ import 'package:network_manager/network_manager.dart';
 class MockDataConnectionChecker extends Mock implements DataConnectionChecker {}
 
 void main() {
-  NetworkManager networkInfo;
-  MockDataConnectionChecker dataConnectionChecker;
+  late NetworkManager networkInfo;
+  late MockDataConnectionChecker dataConnectionChecker;
 
   setUp(() {
     dataConnectionChecker = MockDataConnectionChecker();
@@ -22,14 +22,14 @@ void main() {
         final tIsConnected = Future.value(true);
 
         when(
-          dataConnectionChecker.hasConnection,
+          () => dataConnectionChecker.hasConnection,
         ).thenAnswer((_) => tIsConnected);
 
         // side effects -> collect the result to test
         final result = networkInfo.isConnected;
 
         // expectations -> compare result to expected value
-        verify(dataConnectionChecker.hasConnection);
+        verify(() => dataConnectionChecker.hasConnection);
         expect(result, tIsConnected);
       },
     );
